+++ tests/cli/test_datasets.py
@@ -22,6 +22,7 @@ from __future__ import absolute_import, print_function
 import json
 import os
 import shutil
+import textwrap
 from pathlib import Path
 
 import pytest
@@ -34,6 +35,7 @@ from renku.core.management.config import RENKU_HOME
 from renku.core.management.datasets import DatasetsApiMixin
 from renku.core.management.repository import DEFAULT_DATA_DIR as DATA_DIR
 from renku.core.models.refs import LinkReference
+from tests.utils import assert_dataset_is_mutated
 
 
 def test_datasets_create_clean(runner, project, client):
@@ -288,7 +290,7 @@ def test_dataset_url_in_different_domain(runner, client):
 def test_datasets_list_empty(output_format, runner, project):
     """Test listing without datasets."""
     format_option = "--format={0}".format(output_format)
-    result = runner.invoke(cli, ["dataset", format_option])
+    result = runner.invoke(cli, ["dataset", "ls", format_option])
     assert 0 == result.exit_code
 
 
@@ -300,11 +302,11 @@ def test_datasets_list_non_empty(output_format, runner, project):
     assert 0 == result.exit_code
     assert "OK" in result.output
 
-    result = runner.invoke(cli, ["dataset", format_option])
+    result = runner.invoke(cli, ["dataset", "ls", format_option])
     assert 0 == result.exit_code
     assert "my-dataset" in result.output
 
-    result = runner.invoke(cli, ["dataset", "--revision=HEAD~1", format_option])
+    result = runner.invoke(cli, ["dataset", "ls", "--revision=HEAD~1", format_option])
     assert 0 == result.exit_code
     assert "my-dataset" not in result.output
 
@@ -324,7 +326,7 @@ def test_datasets_list_with_columns(runner, project, columns, headers, values):
     )
     assert 0 == result.exit_code
 
-    result = runner.invoke(cli, ["dataset", "--columns", columns])
+    result = runner.invoke(cli, ["dataset", "ls", "--columns", columns])
     assert 0 == result.exit_code
     assert headers == result.output.split("\n").pop(0).split()
     for value in values:
@@ -334,7 +336,9 @@ def test_datasets_list_with_columns(runner, project, columns, headers, values):
 @pytest.mark.parametrize("column", DATASETS_COLUMNS.keys())
 def test_datasets_list_columns_correctly(runner, project, column):
     """Test dataset listing only shows requested columns."""
-    result = runner.invoke(cli, ["dataset", "--columns", column])
+    assert 0 == runner.invoke(cli, ["dataset", "create", "test"]).exit_code
+
+    result = runner.invoke(cli, ["dataset", "ls", "--columns", column])
     assert 0 == result.exit_code
     header = result.output.split("\n").pop(0)
     name, display_name = DATASETS_COLUMNS[column]
@@ -345,9 +349,24 @@ def test_datasets_list_columns_correctly(runner, project, column):
 @pytest.mark.parametrize("columns", ["invalid", "id,invalid"])
 def test_datasets_list_invalid_column(runner, project, columns):
     """Test dataset listing invalid column name."""
-    result = runner.invoke(cli, ["dataset", "--columns", columns])
+    result = runner.invoke(cli, ["dataset", "ls", "--columns", columns])
     assert 2 == result.exit_code
     assert 'Invalid column name: "invalid".' in result.output
+
+
+def test_datasets_list_description(runner, project):
+    """Test dataset description listing."""
+    description = "Very long description. " * 100
+    assert 0 == runner.invoke(cli, ["dataset", "create", "test", "-d", description]).exit_code
+
+    short_description = textwrap.wrap(description, width=64, max_lines=2)[0]
+
+    result = runner.invoke(cli, ["dataset", "ls", "--columns=name,description"])
+
+    assert 0 == result.exit_code
+    line = next(line for line in result.output.split("\n") if "test" in line)
+    assert short_description in line
+    assert description[: len(short_description) + 1] not in line
 
 
 def test_add_and_create_dataset(directory_tree, runner, project, client, subdirectory):
@@ -408,6 +427,26 @@ def test_add_to_dirty_repo(directory_tree, runner, project, client):
 
     assert client.repo.is_dirty()
     assert ["untracked"] == client.repo.untracked_files
+
+
+def test_add_unicode_file(tmpdir, runner, project, client):
+    """Test adding files with unicode special characters in their names."""
+    # create a dataset
+    result = runner.invoke(cli, ["dataset", "create", "my-dataset"])
+    assert 0 == result.exit_code
+    assert "OK" in result.output
+
+    filename = "filéàèû爱ಠ_ಠ.txt"
+    new_file = tmpdir.join(filename)
+    new_file.write(str("test"))
+
+    # add data
+    result = runner.invoke(cli, ["dataset", "add", "my-dataset", str(new_file)],)
+    assert 0 == result.exit_code
+
+    result = runner.invoke(cli, ["log", "--format", "json-ld", "--strict", f"data/my-dataset/{filename}"])
+    assert 0 == result.exit_code
+    assert filename in result.output.encode("latin1").decode("unicode-escape")
 
 
 def test_multiple_file_to_dataset(tmpdir, runner, project, client):
@@ -600,6 +639,40 @@ def test_datasets_ls_files_check_exit_code(output_format, runner, project):
     format_option = "--format={0}".format(output_format)
     result = runner.invoke(cli, ["dataset", "ls-files", format_option])
     assert 0 == result.exit_code
+
+
+def test_datasets_ls_files_lfs(tmpdir, large_file, runner, project):
+    """Test file listing lfs status."""
+    # NOTE: create a dataset
+    result = runner.invoke(cli, ["dataset", "create", "my-dataset"])
+    assert 0 == result.exit_code
+    assert "OK" in result.output
+
+    # NOTE: create some data
+    paths = []
+
+    new_file = tmpdir.join("file_1")
+    new_file.write(str(1))
+    paths.append(str(new_file))
+
+    paths.append(str(large_file))
+
+    # NOTE: add data to dataset
+    result = runner.invoke(cli, ["dataset", "add", "my-dataset"] + paths, catch_exceptions=False,)
+    assert 0 == result.exit_code
+
+    # NOTE: check files
+    result = runner.invoke(cli, ["dataset", "ls-files"])
+    assert 0 == result.exit_code
+
+    lines = result.output.split("\n")
+    file1_entry = next(line for line in lines if "file_1" in line)
+    file2_entry = next(line for line in lines if large_file.name in line)
+
+    assert file1_entry
+    assert file2_entry
+    assert not file1_entry.endswith("*")
+    assert file2_entry.endswith("*")
 
 
 @pytest.mark.parametrize("column", DATASET_FILES_COLUMNS.keys())
@@ -817,61 +890,29 @@ def test_dataset_unlink_file(tmpdir, runner, client, subdirectory):
         assert new_file.basename not in [file_.path.name for file_ in dataset.files]
 
 
-def test_dataset_1rm(tmpdir, runner, project, client, subdirectory):
+def test_dataset_rm(runner, client, directory_tree, subdirectory):
     """Test removal of a dataset."""
-    # try to delete non existing dataset
-    result = runner.invoke(cli, ["dataset", "rm"])
-    assert 2 == result.exit_code
+    assert 0 == runner.invoke(cli, ["dataset", "add", "--create", "my-dataset", str(directory_tree)]).exit_code
 
-    result = runner.invoke(cli, ["dataset", "rm", "does-not-exist"])
-    assert 2 == result.exit_code
+    dataset = client.load_dataset("my-dataset")
+    assert client.load_dataset("my-dataset")
+    assert (client.path / dataset.path).exists()
 
-    # create a dataset
-    result = runner.invoke(cli, ["dataset", "create", "my-dataset"])
-    assert 0 == result.exit_code
-    assert "OK" in result.output
-
-    # create some data
-    paths = []
-    for i in range(3):
-        new_file = tmpdir.join("file_{0}".format(i))
-        new_file.write(str(i))
-        paths.append(str(new_file))
-
-    # add data to dataset
-    result = runner.invoke(cli, ["dataset", "add", "my-dataset"] + paths, catch_exceptions=False,)
-    assert 0 == result.exit_code
-
-    # try to delete a non empty dataset
     result = runner.invoke(cli, ["dataset", "rm", "my-dataset"])
-    assert 0 == result.exit_code
 
-    # check output
+    assert 0 == result.exit_code, result.output
     assert "OK" in result.output
     assert not client.load_dataset("my-dataset")
+    assert not (client.path / dataset.path).exists()
 
     result = runner.invoke(cli, ["doctor"], catch_exceptions=False)
     assert 0 == result.exit_code
 
 
-def test_dataset_rm_commit(tmpdir, runner, project, client):
-    """Test removal of a dataset repository state."""
-    # create a dataset
-    result = runner.invoke(cli, ["dataset", "create", "my-dataset"])
-    assert 0 == result.exit_code
-    assert "OK" in result.output
-
-    # try to delete a non empty dataset
-    result = runner.invoke(cli, ["dataset", "rm", "my-dataset"])
-    assert 0 == result.exit_code
-
-    # check output
-    assert "OK" in result.output
-    assert not client.load_dataset("my-dataset")
-
-    # Dirty repository check.
-    result = runner.invoke(cli, ["status"])
-    assert 0 == result.exit_code
+def test_dataset_rm_failure(runner, client):
+    """Test errors in removal of a dataset."""
+    assert 2 == runner.invoke(cli, ["dataset", "rm"]).exit_code
+    assert 1 == runner.invoke(cli, ["dataset", "rm", "does-not-exist"]).exit_code
 
 
 def test_dataset_overwrite_no_confirm(runner, project):
@@ -911,7 +952,7 @@ def test_dataset_edit(runner, client, project, dirty, subdirectory):
     dataset = client.load_dataset("dataset")
     assert " new description " == dataset.description
     assert "original title" == dataset.title
-    assert {creator1, creator2} == {c.full_identity for c in dataset.creators}
+    assert {creator1, creator2}.issubset({c.full_identity for c in dataset.creators})
 
     result = runner.invoke(cli, ["dataset", "edit", "dataset", "-t", " new title "], catch_exceptions=False)
     assert 0 == result.exit_code
@@ -926,7 +967,7 @@ def test_dataset_edit(runner, client, project, dirty, subdirectory):
     dataset = client.load_dataset("dataset")
     assert " new description " == dataset.description
     assert "new title" == dataset.title
-    assert {creator1, creator2} == {c.full_identity for c in dataset.creators}
+    assert {creator1, creator2}.issubset({c.full_identity for c in dataset.creators})
     assert {"keyword-2", "keyword-3"} == set(dataset.keywords)
 
 
@@ -973,7 +1014,7 @@ def test_dataset_provider_resolution_dataverse(doi_responses, uri):
     assert type(provider) is DataverseProvider
 
 
-def test_dataset_tag(tmpdir, runner, project, subdirectory):
+def test_dataset_tag(tmpdir, runner, project, client, subdirectory):
     result = runner.invoke(cli, ["dataset", "create", "my-dataset"])
     assert 0 == result.exit_code
     assert "OK" in result.output
@@ -1476,7 +1517,7 @@ def test_workflow_with_external_file(runner, client, directory_tree, project, ru
     result = runner.invoke(cli, ["status"])
     assert 1 == result.exit_code
 
-    assert 0 == run(args=("update",))
+    assert 0 == run(args=("update", "--all",))
     result = runner.invoke(cli, ["status"])
     assert 0 == result.exit_code
 
@@ -1485,3 +1526,83 @@ def test_workflow_with_external_file(runner, client, directory_tree, project, ru
 
     attributes = (client.path / ".gitattributes").read_text().split()
     assert "data/output.txt" in attributes
+
+
+@pytest.mark.parametrize(
+    "args", [["dataset", "create", "my-data"], ["dataset", "add", "--create", "my-data", "README.md"]]
+)
+def test_immutability_at_creation(runner, client, args):
+    """Test first dataset's ID is the same as metadata directory."""
+    assert 0 == runner.invoke(cli, args).exit_code
+
+    dataset = client.load_dataset("my-data")
+    assert str(dataset.path).endswith(dataset.identifier)
+
+
+def test_immutability_for_files(directory_tree, runner, client):
+    """Test dataset's ID changes after a change to dataset files."""
+    assert 0 == runner.invoke(cli, ["dataset", "create", "my-data"]).exit_code
+
+    old_dataset = client.load_dataset("my-data")
+
+    # Add some files
+    assert 0 == runner.invoke(cli, ["dataset", "add", "my-data", str(directory_tree)]).exit_code
+
+    dataset = client.load_dataset("my-data")
+    assert_dataset_is_mutated(old=old_dataset, new=dataset)
+    old_dataset = dataset
+
+    # Remove some files
+    assert 0 == runner.invoke(cli, ["dataset", "unlink", "my-data", "-I", "file1", "--yes"]).exit_code
+
+    dataset = client.load_dataset("my-data")
+    assert_dataset_is_mutated(old=old_dataset, new=dataset)
+
+
+def test_immutability_after_edit(runner, client):
+    """Test dataset's ID changes after editing a dataset."""
+    assert 0 == runner.invoke(cli, ["dataset", "create", "my-data"]).exit_code
+
+    old_dataset = client.load_dataset("my-data")
+
+    assert 0 == runner.invoke(cli, ["dataset", "edit", "my-data", "-k", "new-data"]).exit_code
+
+    dataset = client.load_dataset("my-data")
+    assert_dataset_is_mutated(old=old_dataset, new=dataset)
+
+
+def test_immutability_for_tags(runner, client):
+    """Test dataset's ID does NOT changes after a change to dataset tags."""
+    assert 0 == runner.invoke(cli, ["dataset", "create", "my-data"]).exit_code
+
+    old_dataset = client.load_dataset("my-data")
+
+    # Add a tag
+    assert 0 == runner.invoke(cli, ["dataset", "tag", "my-data", "new-tag"]).exit_code
+
+    dataset = client.load_dataset("my-data")
+    assert old_dataset._id == dataset._id
+    old_dataset = dataset
+
+    # Remove a tag
+    assert 0 == runner.invoke(cli, ["dataset", "rm-tags", "my-data", "new-tag"]).exit_code
+
+    dataset = client.load_dataset("my-data")
+    assert old_dataset._id == dataset._id
+
+
+def test_immutability_after_remove(directory_tree, runner, client):
+    """Test dataset is mutated one final time when it is removed."""
+    assert 0 == runner.invoke(cli, ["dataset", "create", "my-data"]).exit_code
+
+    old_dataset = client.load_dataset("my-data")
+
+    assert 0 == runner.invoke(cli, ["dataset", "rm", "my-data"]).exit_code
+
+    assert client.load_dataset("my-data") is None
+
+    # Checkout previous commit that has dataset's final version
+    client.repo.git.checkout("HEAD~")
+
+    dataset = client.load_dataset("my-data")
+    assert_dataset_is_mutated(old=old_dataset, new=dataset)