+++ tests/utils.py
@@ -55,3 +55,22 @@ def make_dataset_add_payload(project_id, urls, name=None):
         "force": False,
         "files": files,
     }
+
+
+def assert_dataset_is_mutated(old, new, mutator=None):
+    """Check metadata is updated correctly after dataset mutation."""
+    assert old._id != new._id
+    assert old.identifier != new.identifier
+    assert old.path == new.path
+    assert old._id == new.derived_from.url["@id"]
+    assert old.date_created != new.date_created
+    assert new.same_as is None
+    assert new.date_published is None
+    assert new.identifier in new._id
+    assert new.identifier in new._label
+    assert new.identifier in new.url
+
+    if mutator:
+        old_creators = {c.email for c in old.creators}
+        new_creators = {c.email for c in new.creators}
+        assert new_creators == old_creators | {mutator.email}